requirements.txt
langchain
langgraph
langchain-openai
openai
pandas
numpy
scikit-learn
matplotlib
seaborn
tqdm
jupyter
ipython
requests
python-dotenv
streamlit
